+++ src/main/java/net/silentchaos512/gear/api/material/IMaterialInstance.java
@@ -3,6 +3,7 @@ package net.silentchaos512.gear.api.material;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
 import net.silentchaos512.gear.api.parts.MaterialGrade;
 import net.silentchaos512.gear.api.parts.PartType;
 
@@ -24,5 +25,11 @@ public interface IMaterialInstance {
 
     default int getColor(PartType partType) {
         return getColor(partType, ItemStack.EMPTY);
+    }
+
+    ITextComponent getDisplayName(PartType partType, ItemStack gear);
+
+    default ITextComponent getDisplayName(PartType partType) {
+        return getDisplayName(partType, ItemStack.EMPTY);
     }
 }