+++ src/main/java/net/silentchaos512/gear/parts/type/CompoundPart.java
@@ -1,7 +1,11 @@
 package net.silentchaos512.gear.parts.type;
 
+import com.google.gson.JsonObject;
 import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketBuffer;
+import net.minecraft.util.JSONUtils;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
 import net.minecraftforge.common.MinecraftForge;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.event.GetStatModifierEvent;
@@ -16,8 +20,11 @@ import net.silentchaos512.gear.item.CompoundPartItem;
 import net.silentchaos512.gear.parts.AbstractGearPart;
 import net.silentchaos512.gear.parts.PartData;
 import net.silentchaos512.gear.parts.PartPositions;
+import net.silentchaos512.gear.parts.PartTextureType;
 import net.silentchaos512.gear.util.TraitHelper;
+import net.silentchaos512.utils.EnumUtils;
 
+import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -25,21 +32,24 @@ import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public final class CompoundRodPart extends AbstractGearPart {
-    public static final IPartSerializer<CompoundRodPart> SERIALIZER = new Serializer(SilentGear.getId("compound_rod"), CompoundRodPart::new);
+public class CompoundPart extends AbstractGearPart {
+    public static final Serializer SERIALIZER = new Serializer(SilentGear.getId("compound_part"), CompoundPart::new);
 
-    public CompoundRodPart(ResourceLocation name) {
+    private PartType partType;
+    private IPartPosition partPosition;
+
+    public CompoundPart(ResourceLocation name) {
         super(name);
     }
 
     @Override
     public PartType getType() {
-        return PartType.ROD;
+        return this.partType;
     }
 
     @Override
     public IPartPosition getPartPosition() {
-        return PartPositions.ROD;
+        return this.partPosition;
     }
 
     @Override
@@ -48,8 +58,23 @@ public final class CompoundRodPart extends AbstractGearPart {
     }
 
     @Override
+    public PartTextureType getLiteTexture(PartData part, ItemStack gear) {
+        MaterialInstance material = CompoundPartItem.getPrimaryMaterial(part.getCraftingItem());
+        if (material != null) {
+            return material.getMaterial().getTexture(partType, gear);
+        }
+        return super.getLiteTexture(part, gear);
+    }
+
+    @Override
     public int getColor(PartData part, ItemStack gear, int animationFrame) {
         return CompoundPartItem.getColor(part.getCraftingItem());
+    }
+
+    @Override
+    public ITextComponent getDisplayName(@Nullable PartData part, ItemStack gear) {
+        //noinspection ConstantConditions
+        return part != null ? part.getCraftingItem().getDisplayName() : super.getDisplayName(part, gear);
     }
 
     @Override
@@ -64,7 +89,7 @@ public final class CompoundRodPart extends AbstractGearPart {
         // Get the rod materials and all the stat modifiers they provide for this stat
         Collection<MaterialInstance> materials = CompoundPartItem.getMaterials(stack);
         Collection<StatInstance> statMods = materials.stream()
-                .flatMap(m -> m.getStatModifiers(stat, this.getType(), gear).stream())
+                .flatMap(m -> m.getStatModifiers(stat, this.partType, gear).stream())
                 .collect(Collectors.toList());
 
         // Average together all modifiers of the same op. This makes things like rods with varying
@@ -91,9 +116,40 @@ public final class CompoundRodPart extends AbstractGearPart {
         return ret;
     }
 
-    private static class Serializer extends AbstractGearPart.Serializer<CompoundRodPart> {
-        public Serializer(ResourceLocation serializerId, Function<ResourceLocation, CompoundRodPart> function) {
+    private static class Serializer extends AbstractGearPart.Serializer<CompoundPart> {
+        Serializer(ResourceLocation serializerId, Function<ResourceLocation, CompoundPart> function) {
             super(serializerId, function);
+        }
+
+        @Override
+        public CompoundPart read(ResourceLocation id, JsonObject json) {
+            CompoundPart part = super.read(id, json);
+            part.partType = PartType.get(new ResourceLocation(JSONUtils.getString(json, "part_type")));
+            // FIXME
+            part.partPosition = EnumUtils.byName(JSONUtils.getString(json, "part_position"), PartPositions.ANY);
+            return part;
+        }
+
+        @Override
+        public CompoundPart read(ResourceLocation id, PacketBuffer buffer) {
+            CompoundPart part = super.read(id, buffer);
+            part.partType = PartType.get(buffer.readResourceLocation());
+            // FIXME
+            part.partPosition = EnumUtils.byName(buffer.readString(), PartPositions.ANY);
+            return part;
+        }
+
+        @Override
+        public void write(PacketBuffer buffer, CompoundPart part) {
+            super.write(buffer, part);
+            buffer.writeResourceLocation(part.partType.getName());
+            // FIXME
+            for (PartPositions pos : PartPositions.values()) {
+                if (pos == part.partPosition) {
+                    buffer.writeString(pos.name());
+                    break;
+                }
+            }
         }
     }
 }